+++ frontend/src/pages/admin/manageOrder/ManageOrder.tsx
@@ -1,36 +1,39 @@
 import { useState, useEffect } from "react";
-import Header from "../../components/common/Header/Header";
-import Sidebar from "../../components/common/Sidebar/Sidebar";
-import styles from "./ManageUser.module.scss";
+import Header from "../../../components/common/Header/Header";
+import Sidebar from "../../../components/common/Sidebar/Sidebar";
+import styles from "./ManageOrder.module.scss";
 import { FiSearch } from "react-icons/fi";
-import { memberApi } from "../../api/management/memberApi";
-import { User, UserSearchType } from "../../types/user";
+import { orderApi } from "../../../api/admin/orderApi";
+import {
+  getOrderStatusText,
+  getOrderStatusClass,
+} from "../../../utils/orderUtils";
+import { Order, OrderSearchType } from "../../../types/admin/order";
 
-const ManageUser = () => {
+const ManageOrder = () => {
   // 상태 관리
   const [tempSearchTerm, setTempSearchTerm] = useState(""); // 검색어 임시 저장
-  const [searchType, setSearchType] = useState<UserSearchType>("id"); // 검색 타입 (아이디/이름)
+  const [searchType, setSearchType] = useState<OrderSearchType>("orderId"); // 검색 타입 (주문번호/주문날짜 등)
   const [currentPage, setCurrentPage] = useState(1); // 현재 페이지 번호
+  const [orders, setOrders] = useState<Order[]>([]); // 주문 목록
   const [loading, setLoading] = useState(true); // 로딩 상태
-  const [users, setUsers] = useState<User[]>([]); // 사용자 목록
-  const itemsPerPage = 10; // 페이지당 표시할 사용자 수
+  const itemsPerPage = 10; // 페이지당 표시할 주문 수
 
-  // API에서 사용자 데이터 가져오기
-  const fetchUsers = async () => {
-    try {
-      setLoading(true);
-      const data = await memberApi.getMemberList();
-      setUsers(data);
-    } catch (error) {
-      console.error("사용자 데이터를 가져오는 중 오류 발생:", error);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  // 컴포넌트 마운트 시 사용자 데이터 로드
+  // API에서 주문 데이터 가져오기
   useEffect(() => {
-    fetchUsers();
+    const fetchOrders = async () => {
+      try {
+        setLoading(true);
+        const response = await orderApi.getOrderList();
+        setOrders(response);
+      } catch (error) {
+        console.error("주문 데이터를 가져오는 중 오류 발생:", error);
+      } finally {
+        setLoading(false);
+      }
+    };
+
+    fetchOrders();
   }, []);
 
   // 검색어 입력 처리
@@ -38,23 +41,29 @@ const ManageUser = () => {
     setTempSearchTerm(e.target.value);
   };
 
-  // 사용자 검색 실행
+  // 주문 검색 실행
   const handleSearchSubmit = async () => {
     setCurrentPage(1);
-
     try {
       setLoading(true);
-
-      // 검색어가 있을 때만 검색 API 호출
       const searchParams = {
-        memberId: searchType === "id" ? tempSearchTerm.trim() : undefined,
-        memberName: searchType === "name" ? tempSearchTerm.trim() : undefined,
+        orderId:
+          searchType === "orderId" ? parseInt(tempSearchTerm) : undefined,
+        productName: searchType === "product" ? tempSearchTerm : undefined,
+        memberId: searchType === "memberId" ? tempSearchTerm : undefined,
+        paymentName: searchType === "payment" ? tempSearchTerm : undefined,
+        orderDt: searchType === "orderDate" ? tempSearchTerm : undefined,
+        orderState:
+          searchType === "orderStatus" ? parseInt(tempSearchTerm) : undefined,
       };
 
-      const data = await memberApi.searchMember(searchParams);
-      setUsers(data);
+      const response =
+        tempSearchTerm === ""
+          ? await orderApi.getOrderList()
+          : await orderApi.searchOrder(searchParams);
+      setOrders(response);
     } catch (error) {
-      console.error("사용자 검색 중 오류 발생:", error);
+      console.error("주문 데이터를 가져오는 중 오류 발생:", error);
     } finally {
       setLoading(false);
     }
@@ -68,16 +77,16 @@ const ManageUser = () => {
   };
 
   // 검색 타입 변경 처리
-  const handleSearchTypeChange = (type: UserSearchType) => {
+  const handleSearchTypeChange = (type: OrderSearchType) => {
     setSearchType(type);
     setCurrentPage(1);
   };
 
   // 페이지네이션 관련 로직
-  const totalPages = Math.ceil(users.length / itemsPerPage);
+  const totalPages = Math.ceil(orders.length / itemsPerPage);
   const indexOfLastItem = currentPage * itemsPerPage;
   const indexOfFirstItem = indexOfLastItem - itemsPerPage;
-  const currentItems = users.slice(indexOfFirstItem, indexOfLastItem);
+  const currentItems = orders.slice(indexOfFirstItem, indexOfLastItem);
 
   // 페이지 변경 처리
   const handlePageChange = (pageNumber: number) => {
@@ -112,25 +121,52 @@ const ManageUser = () => {
         <div className={styles.mainContent}>
           {/* 페이지 헤더 */}
           <div className={styles.pageHeader}>
-            <h1>사용자 관리</h1>
+            <h1>주문 관리</h1>
           </div>
 
           {/* 검색 섹션 */}
           <div className={styles.filterSection}>
             <div className={styles.searchContainer}>
-              {/* 검색 타입 버튼 */}
+              {/* 검색 타입 버튼 - 첫 번째 줄 */}
               <div className={styles.searchTypeButtons}>
                 <button
-                  className={`${styles.searchTypeButton} ${searchType === "id" ? styles.active : ""}`}
-                  onClick={() => handleSearchTypeChange("id")}
+                  className={`${styles.searchTypeButton} ${searchType === "orderId" ? styles.active : ""}`}
+                  onClick={() => handleSearchTypeChange("orderId")}
                 >
-                  아이디
+                  주문번호
                 </button>
                 <button
-                  className={`${styles.searchTypeButton} ${searchType === "name" ? styles.active : ""}`}
-                  onClick={() => handleSearchTypeChange("name")}
+                  className={`${styles.searchTypeButton} ${searchType === "orderDate" ? styles.active : ""}`}
+                  onClick={() => handleSearchTypeChange("orderDate")}
                 >
-                  이름
+                  주문날짜
+                </button>
+                <button
+                  className={`${styles.searchTypeButton} ${searchType === "orderStatus" ? styles.active : ""}`}
+                  onClick={() => handleSearchTypeChange("orderStatus")}
+                >
+                  주문상태
+                </button>
+              </div>
+              {/* 검색 타입 버튼 - 두 번째 줄 */}
+              <div className={styles.searchTypeButtons}>
+                <button
+                  className={`${styles.searchTypeButton} ${searchType === "memberId" ? styles.active : ""}`}
+                  onClick={() => handleSearchTypeChange("memberId")}
+                >
+                  주문자ID
+                </button>
+                <button
+                  className={`${styles.searchTypeButton} ${searchType === "product" ? styles.active : ""}`}
+                  onClick={() => handleSearchTypeChange("product")}
+                >
+                  물품명
+                </button>
+                <button
+                  className={`${styles.searchTypeButton} ${searchType === "payment" ? styles.active : ""}`}
+                  onClick={() => handleSearchTypeChange("payment")}
+                >
+                  결제수단
                 </button>
               </div>
               {/* 검색 입력창 */}
@@ -153,7 +189,7 @@ const ManageUser = () => {
             </div>
           </div>
 
-          {/* 사용자 목록 테이블 */}
+          {/* 주문 목록 테이블 */}
           <div className={styles.tableContainer}>
             {loading ? (
               // 로딩 상태 표시
@@ -169,32 +205,41 @@ const ManageUser = () => {
                 </p>
               </div>
             ) : (
-              // 사용자 목록 테이블
+              // 주문 목록 테이블
               <div className={styles.tableWrapper}>
-                <table className={styles.userTable}>
+                <table className={styles.orderTable}>
                   <thead>
                     <tr>
-                      <th>ID</th>
-                      <th>이름</th>
-                      <th>전화번호</th>
-                      <th>이메일</th>
-                      <th>생년월일</th>
-                      <th>권한</th>
+                      <th>주문번호</th>
+                      <th>여행 상품명</th>
+                      <th>물품명</th>
+                      <th>주문자 ID</th>
+                      <th>결제수단</th>
+                      <th>여행 상품 수량</th>
+                      <th>물품 수량</th>
+                      <th>주문 일자</th>
+                      <th>주문 상태</th>
                     </tr>
                   </thead>
                   <tbody>
-                    {currentItems.map(user => (
-                      <tr key={user.memberId}>
-                        <td>{user.memberId}</td>
-                        <td>{user.memberName}</td>
-                        <td>{user.memberPhone}</td>
-                        <td>{user.memberEmail}</td>
-                        <td>{user.memberBirth}</td>
+                    {currentItems.map(order => (
+                      <tr key={order.orderId}>
+                        <td>{order.orderId}</td>
+                        <td>{order.travelName || "-"}</td>
+                        <td>{order.productName || "-"}</td>
+                        <td>{order.memberId}</td>
+                        <td>{order.paymentName}</td>
+                        <td>{order.orderTravelAmount}</td>
+                        <td>{order.orderProductAmount}</td>
+                        <td>{order.orderDt}</td>
                         <td>
                           <span
-                            className={`${styles.roleTag} ${user.authName === "ADMIN" ? styles.adminRole : styles.guestRole}`}
+                            className={`${styles.statusTag} ${getOrderStatusClass(
+                              order.orderState,
+                              styles,
+                            )}`}
                           >
-                            {user.authName === "ADMIN" ? "관리자" : "사용자"}
+                            {getOrderStatusText(order.orderState)}
                           </span>
                         </td>
                       </tr>
@@ -243,4 +288,4 @@ const ManageUser = () => {
   );
 };
 
-export default ManageUser;
+export default ManageOrder;
