+++ frontend/src/pages/admin/manageProduct/ManageProduct.module.scss
