+++ frontend/src/types/admin/dashboard.ts
