+++ frontend/src/types/admin/product.ts
