+++ frontend/src/utils/transformProduct.ts
@@ -1,5 +1,4 @@
-import { ApiProduct } from "../types/product";
-import { Product } from "../types/product";
+import { ApiProduct, Product } from "../types/admin/product";
 
 // API 응답 데이터를 Product 타입으로 변환하는 함수
 export const transformApiProduct = (item: ApiProduct): Product => ({
